python-service/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.8.2
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML and Transformers
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2

# Monitoring and logging
prometheus-client==0.19.0

